refactor: change single radio api
