+++ packages/blade/src/components/Radio/Radio.stories.tsx
@@ -1,42 +1,69 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
-import React from 'react';
 import type { ComponentStory, Meta } from '@storybook/react';
-import { Button, Text } from '..';
-import type { RadioProps } from './Radio';
+import React from 'react';
+import { Text } from '../Typography';
+import type { RadioGroupProps } from './RadioGroup/RadioGroup';
+import { RadioGroup as RadioGroupComponent } from './RadioGroup/RadioGroup';
 import { Radio as RadioComponent } from './Radio';
-import Box from '~components/Box';
 
 export default {
   title: 'Components/Radio/Radio',
-  component: RadioComponent,
+  component: RadioGroupComponent,
   args: {
-    defaultChecked: undefined,
+    label: 'Radio example',
+    helpText: undefined,
+    isDisabled: false,
+    neccessityIndicator: 'none',
+    labelPosition: undefined,
     validationState: undefined,
-    isChecked: undefined,
-    isDisabled: undefined,
-    isIndeterminate: undefined,
-    isRequired: undefined,
+    errorText: undefined,
     name: undefined,
+    defaultValue: undefined,
     onChange: undefined,
     value: undefined,
-    helpText: undefined,
-    errorText: undefined,
-    labelPosition: 'top',
-    children: 'Toggle checkbox',
   },
-} as Meta<RadioProps>;
+  argTypes: {
+    value: {
+      options: ['apple', 'mango', 'orange'],
+      control: {
+        type: 'select',
+      },
+    },
+    defaultValue: {
+      options: ['apple', 'mango', 'orange'],
+      control: {
+        type: 'select',
+      },
+    },
+  },
+} as Meta<RadioGroupProps>;
 
-const RadioTemplate: ComponentStory<typeof RadioComponent> = ({ children, ...args }) => {
-  return <RadioComponent {...args}>{children}</RadioComponent>;
+const RadioTemplate: ComponentStory<typeof RadioGroupComponent> = ({ children, ...args }) => {
+  return (
+    <RadioGroupComponent {...args}>
+      <RadioComponent value="apple">Apple</RadioComponent>
+      <RadioComponent isDisabled value="mango">
+        Mango
+      </RadioComponent>
+      <RadioComponent value="orange">Orange</RadioComponent>
+    </RadioGroupComponent>
+  );
 };
 
 export const Default = RadioTemplate.bind({});
 Default.storyName = 'Default';
 
-export const Checked = RadioTemplate.bind({});
-Checked.storyName = 'Checked';
-Checked.args = {
-  isChecked: true,
+export const HelpText = RadioTemplate.bind({});
+HelpText.storyName = 'HelpText';
+HelpText.args = {
+  helpText: 'RadioGroup help text',
+};
+
+export const ErrorText = RadioTemplate.bind({});
+ErrorText.storyName = 'ErrorText';
+ErrorText.args = {
+  validationState: 'error',
+  errorText: 'RadioGroup help text',
 };
 
 export const Disabled = RadioTemplate.bind({});
@@ -45,37 +72,90 @@ Disabled.args = {
   isDisabled: true,
 };
 
-export const Invalid = RadioTemplate.bind({});
-Invalid.storyName = 'Invalid';
-Invalid.args = {
-  validationState: 'error',
+export const Optional = RadioTemplate.bind({});
+Optional.storyName = 'Optional';
+Optional.args = {
+  neccessityIndicator: 'optional',
 };
 
-const ControlledAndUncontrolledComp = () => {
-  const [checked, setChecked] = React.useState(false);
+export const Required = RadioTemplate.bind({});
+Required.storyName = 'Required';
+Required.args = {
+  neccessityIndicator: 'required',
+};
+
+export const LabelPositionLeft = RadioTemplate.bind({});
+LabelPositionLeft.storyName = 'LabelPositionLeft';
+LabelPositionLeft.args = {
+  labelPosition: 'left',
+};
+
+export const KitchenSink = (): React.ReactElement => {
+  const [selected, setSelected] = React.useState('orange');
 
   return (
     <>
-      <RadioComponent name="fruits">Uncontrolled</RadioComponent>
-      <Text>&nbsp;</Text>
-      <RadioComponent
-        isChecked={checked}
-        onChange={({ isChecked }) => {
-          setChecked(isChecked);
-        }}
-        name="fruits"
+      <RadioGroupComponent
+        helpText="Select atleast one"
+        label="Uncontrolled"
+        defaultValue="orange"
+        onChange={(e) => console.log(e)}
       >
-        Controlled
-      </RadioComponent>
-      <Text>Checked: {checked ? 'True' : 'False'}</Text>
-      <Box marginBottom="spacing.2" />
-      <Button size="small" onClick={() => setChecked((prev) => !prev)}>
-        Toggle checked
-      </Button>
+        <RadioComponent value="apple">Apple</RadioComponent>
+        <RadioComponent value="mango">Mango</RadioComponent>
+        <RadioComponent value="orange">Orange</RadioComponent>
+      </RadioGroupComponent>
+      <Text>&nbsp;</Text>
+      <RadioGroupComponent
+        errorText="Selected atleast one item"
+        helpText={`You selected ${selected}`}
+        label="Controlled"
+        value={selected}
+        onChange={({ value, name }) => {
+          setSelected(value);
+          console.log(name, value);
+        }}
+      >
+        <RadioComponent value="apple">Apple</RadioComponent>
+        <RadioComponent value="mango">Mango</RadioComponent>
+        <RadioComponent value="orange">Orange</RadioComponent>
+      </RadioGroupComponent>
+      <Text>&nbsp;</Text>
+      <RadioGroupComponent
+        neccessityIndicator="required"
+        errorText="Atleast one has to be selected"
+        helpText="Select atleast one"
+        label="Select your fruit"
+      >
+        <RadioComponent value="apple">Apple</RadioComponent>
+        <RadioComponent value="mango">Mango</RadioComponent>
+        <RadioComponent value="orange">Orange</RadioComponent>
+      </RadioGroupComponent>
+      <Text>&nbsp;</Text>
+      <RadioGroupComponent
+        validationState="error"
+        neccessityIndicator="optional"
+        errorText="Atleast one has to be selected"
+        helpText="Select atleast one"
+        label="Select your fruit"
+      >
+        <RadioComponent value="apple">Apple</RadioComponent>
+        <RadioComponent value="mango">Mango</RadioComponent>
+        <RadioComponent value="orange">Orange</RadioComponent>
+      </RadioGroupComponent>
+      <Text>&nbsp;</Text>
+      <RadioGroupComponent
+        labelPosition="left"
+        neccessityIndicator="optional"
+        validationState="error"
+        errorText="This is invalid"
+        helpText="Select atleast one"
+        label="Select your fruit"
+      >
+        <RadioComponent value="apple">Apple</RadioComponent>
+        <RadioComponent value="mango">Mango</RadioComponent>
+        <RadioComponent value="orange">Orange</RadioComponent>
+      </RadioGroupComponent>
     </>
   );
 };
-const _ControlledAndUncontrolled: ComponentStory<typeof RadioComponent> = () => {
-  return <ControlledAndUncontrolledComp />;
-};
-export const ControlledAndUncontrolled = _ControlledAndUncontrolled.bind({});
